+++ src/api/users/users.rs
@@ -142,6 +142,9 @@ pub struct Users<'a> {
     /// These are generally Service Desk users or other GitLab-managed users.
     #[builder(default)]
     exclude_internal: Option<bool>,
+    /// Exclude external users.
+    #[builder(default)]
+    exclude_external: Option<bool>,
     /// Filter uses based on administrator status.
     #[builder(default)]
     admins: Option<bool>,
@@ -216,6 +219,7 @@ impl<'a> Endpoint for Users<'a> {
             .push_opt("without_projects", self.without_projects)
             .push_opt("without_project_bots", self.without_project_bots)
             .push_opt("exclude_internal", self.exclude_internal)
+            .push_opt("exclude_external", self.exclude_external)
             .push_opt("admins", self.admins)
             .push_opt("saml_provider_id", self.saml_provider_id);
 
@@ -541,6 +545,19 @@ mod tests {
         let client = SingleTestClient::new_raw(endpoint, "");
 
         let endpoint = Users::builder().exclude_internal(false).build().unwrap();
+        api::ignore(endpoint).query(&client).unwrap();
+    }
+
+    #[test]
+    fn endpoint_exclude_external() {
+        let endpoint = ExpectedUrl::builder()
+            .endpoint("users")
+            .add_query_params(&[("exclude_external", "false")])
+            .build()
+            .unwrap();
+        let client = SingleTestClient::new_raw(endpoint, "");
+
+        let endpoint = Users::builder().exclude_external(false).build().unwrap();
         api::ignore(endpoint).query(&client).unwrap();
     }