+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   * Support `Any`, `Today`, and `Tomorrow` issue due date filters.
   * Support sorting issue results by their title.
   * Support filtering group issues by their type.
+  * Support filtering out external users when listing users.
 
 ## Changes